phish-web/api/requirements.txt
anyio==3.6.2
beautifulsoup4==4.12.2
boto3==1.26.132
botocore==1.29.132
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.1
filelock==3.12.0
h11==0.14.0
humanize==4.6.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
Levenshtein==0.21.0
nose==1.3.7
nostril @ git+https://github.com/casics/nostril.git@fbc0c91249283a9fbc9036206391ce1138826fd3
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pickle5==0.0.11
plac==1.3.5
pluggy==1.0.0
pydantic==1.10.7
pytest==7.3.1
python-dateutil==2.8.2
python-Levenshtein==0.21.0
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.0.0
requests==2.31.0
requests-file==1.5.1
s3transfer==0.6.1
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
tabulate==0.9.0
tldextract==3.4.4
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
xgboost==1.7.5
